define C++11 advanced exceptions unconditionally
