+++ ChangeLog
@@ -1,3 +1,10 @@
+2025-03-24  Iñaki Ucar  <iucar@fedoraproject.org>
+
+	* inst/include/Rcpp/exceptions.h: Include C+11 variadic template definitions
+	unconditionally
+	* inst/include/Rcpp/exceptions/cpp11/exceptions.h: Removed
+	* inst/include/Rcpp/exceptions/cpp98/exceptions.h: Idem
+
 2025-03-21  Dirk Eddelbuettel  <edd@debian.org>
 
 	* DESCRIPTION (Version, Date): Roll micro version and date
